requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
pycparser==2.19
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.25.3
